refactor(marathon): migrate announcements to feature module
