+++ features/marathon/constants.ts
@@ -1,0 +1,22 @@
+export const announcementItems = [
+  {
+    title: "🚨島島盃｜2025 春季學習馬拉松 即將於1月24日23:59截止",
+    tag: "活動公告",
+    times: "2024 / 12 / 16",
+    content:
+      "![](https://image.daoedu.tw/fe9ce0f8-2034-4801-b62f-a31a8bf58c1f)【島島盃｜2025 春季學習馬拉松 開跑囉！1/24 23:59報名申請截止】\n\n🏃🏻 「為什麼想做的事總是跑不起來？」，別擔心，島島盃引導師、AI、社群陪你跑五個月，跑向屬於自己的終點！\n\n發展興趣、改變生活習慣、上理想的大學、生涯規劃、發起社會行動，每一個生活大小事都是一場學習馬拉松。\n\n然而總會遇到「不知道怎麼計畫」、「好難自律」、「沒有伴」、「資源與人脈有限」、「無限自我質疑」等難題...\n\n「島島盃學習馬拉松」將提供你四大裝備：\n\n![](https://image.daoedu.tw/14ecaa39-9dd1-4f18-be58-6761a30bba3f)\n\n如果你有些想做的計畫，正在等待個契機開始，現在就是時候。\n\n五個月的馬拉松，你將會在計畫過程中「豐富知識經驗、在學習中形塑自我、為生活與社會帶來實際行動」， #完賽不僅全額退費還有機會獲得獎助金。\n\n![](https://image.daoedu.tw/3f53dc57-621b-4103-a1ad-d1fd96dff44b)\n\n島島盃 2025 春季學習馬拉松，將以學習者的自我需求出發設計學習計畫，開啟一趟自我導向學習馬拉松，往哪跑？怎麼跑？跑多快？終點在哪由你決定，島島阿學陪你一起跑。\n\n邀請你一起「為自己重新打造喜歡的學習生活」，把自主學習變成一種生活方式，別讓「等等再開始」成為絆腳石！\n\n【島島盃－2025 春季學習馬拉松】報名資訊\n\n➤報名時間｜即日起～2025年1月24日 23:59止\n\n➤招募對象｜16歲以上學習者皆可報名，優先以高中及大學生為主；任何有意願為自己打造專屬學習旅程的學習者\n\n➤計畫官網與說明｜[https://www.daoedu.tw/learning-marathon#marathon-intro](https://www.daoedu.tw/learning-marathon#marathon-intro)\n\n➤計畫時程｜2025/2/9-2025/7/12\n\n其他優惠折抵活動：\n\n👉【活動一】折價券申請（以下兩種方案皆可使用）：\n\n1.於IG限時動態或FB限時動態貼文分享島島盃－2025 春季學習馬拉松活動：折抵 500 元\n\n2.參與自主學習工作坊暨說明會的夥伴：折抵 500 元\n\n表單連結：[https://forms.gle/9Pfa9Q5d27m1JEpUA](https://forms.gle/9Pfa9Q5d27m1JEpUA)\n\n👉【活動二】團體報名：\n\n團體報名享有更多優惠！只要填寫報名申請的Email信箱我們會幫忙促成團報喔\\~\n\n表單連結：[https://forms.gle/zRwyNKtPPfvM1qbx9](https://forms.gle/zRwyNKtPPfvM1qbx9)\n\n更多關於島島盃學習馬拉松的實用資訊：\n\n1.[【用 AI 好好寫學習計畫】八步驟 Prompt 大公開！](https://www.instagram.com/p/DE6hMytzPDX/?utm_source=ig_web_copy_link\\&igsh=MzRlODBiNWFlZA==)\n\n2.[【六步驟，陪你寫一份自己滿意也可以執行的自主學習計畫】](https://www.instagram.com/p/DEhmX2dhrjz/?utm_source=ig_web_copy_link\\&igsh=MzRlODBiNWFlZA==)\n\n3.[【島島盃｜2025 春季學習馬拉松】專業引導師是誰？](https://www.instagram.com/p/DEAIjbUTTh0/?utm_source=ig_web_copy_link\\&igsh=MzRlODBiNWFlZA==)",
+  },
+  {
+    title: "🚨島島盃－2025 春季學習馬拉松  FAQ 大解密",
+    tag: "FAQ",
+    times: "2025 / 01 / 09",
+    content:
+      "![](https://image.daoedu.tw/58f3c0c7-fde1-430d-aa91-ec09d746c9fe)**#申請截止日可以延長嗎？**\n\n還在猶豫是否報名？對於資格、費用或完賽標準有疑問嗎？別擔心，所有重點已幫你整理好，快來加入這場挑戰學習力與成長的旅程吧！\n\n為了讓大家有更多時間準備學習計畫，考量到期末考與學測，申請截止日已延至 01/24 (五) 23:59。繳交後，在截止前仍可以不斷修改內容，務必把握時間！\n\n**#16歲以下可以報名嗎？**\n\n最近收到16歲以下的國中同學說「真的很想參加，有沒有機會！」經主辦單位討論後，我們希望支持每一位有學習熱情的夥伴，同時考量活動進行方式，16歲以下的國一以上（含國一）學生，若有高度申請動機，且經法定代理人同意者，歡迎寄信給主辦單位。\n\n**#如果申請時一起團報的朋友沒有入圍，還可以享有團報費用嗎？**\n\n可以！只要填寫好團體報名表並經由主辦方確認團報成功，無論其他人是否錄取，你都可以享有團體報名優惠。\n\n**#完賽的定義是什麼？**\n\n完賽不看最終成果，而看你的參與與成長：過程中的反思與改變，對自我或社會的影響，以及未來的行動方向。\n\n完賽的退費標準為：\n\n1️⃣ 諮詢與工作坊請假總時數不超過 5 小時。\n\n2️⃣ 提交每兩週的進度報告。\n\n3️⃣ 參加 7/12 成果發表日。\n\n4️⃣ 完成並上傳成果發表資料、分享至少 3 個學習資源與心得，並完成問卷調查。\n\n**#報名費最低真的只要 1000 嗎？**\n\n是的！我們提供多種折抵方案：\n\n只要透過四人團報（3,000元／人），並完成以下三者，可以最低費用參與：\n\n1.參加任一場線上說明會：折抵 NT 500\n\n3.IG 限時動態 或 FB 貼文按讚並標記 2 位朋友：折抵 NT 500\n\n3.前五名繳交申請報名：折抵 NT 1,000\n\n**#怎麼才算報名成功？**\n\n只要在「核對學習內容」後按下提交申請，並看到成功頁面與收到確認郵件，就代表報名完成囉！\n\n⏳ 時間不等人，立即提交你的學習挑戰計畫吧！\n\n【島島盃－2025 春季學習馬拉松】報名資訊\n\n➤報名時間｜即日起～2025年1月24日 23:59止\n\n➤招募對象｜16歲以上學習者皆可報名，優先以高中及大學生為主；任何有意願為自己打造專屬學習旅程的學習者\n\n➤計畫官網與說明｜[https://www.daoedu.tw/learning-marathon#marathon-intro](https://www.daoedu.tw/learning-marathon#marathon-intro)\n\n➤計畫時程｜2025/2/9-2025/7/12\n\n其他優惠折抵活動：\n\n👉【活動一】折價券申請（以下兩種方案皆可使用）：\n\n1.於IG限時動態或FB限時動態貼文分享島島盃－2025 春季學習馬拉松活動：折抵 500 元\n\n2.參與自主學習工作坊暨說明會的夥伴：折抵 500 元\n\n表單連結：[https://forms.gle/9Pfa9Q5d27m1JEpUA](https://forms.gle/9Pfa9Q5d27m1JEpUA)\n\n👉【活動二】團體報名：\n\n團體報名享有更多優惠！只要填寫報名申請的Email信箱我們會幫忙促成團報喔\\~\n\n表單連結：[https://forms.gle/zRwyNKtPPfvM1qbx9](https://forms.gle/zRwyNKtPPfvM1qbx9)\n\n更多關於島島盃學習馬拉松的實用資訊：\n\n1.[【用 AI 好好寫學習計畫】八步驟 Prompt 大公開！](https://www.instagram.com/p/DE6hMytzPDX/?utm_source=ig_web_copy_link\\&igsh=MzRlODBiNWFlZA==)\n\n2.[【六步驟，陪你寫一份自己滿意也可以執行的自主學習計畫】](https://www.instagram.com/p/DEhmX2dhrjz/?utm_source=ig_web_copy_link\\&igsh=MzRlODBiNWFlZA==)\n\n3.[【島島盃｜2025 春季學習馬拉松】專業引導師是誰？](https://www.instagram.com/p/DEAIjbUTTh0/?utm_source=ig_web_copy_link\\&igsh=MzRlODBiNWFlZA==)",
+  },
+]
+  .map((item, index) => ({
+    ...item,
+    id: `${index + 1}`,
+    author: "島小編",
+  }))
+  .reverse();
