+++ pages/learning-marathon/announcements/index.tsx
@@ -4,8 +4,7 @@ import Nav from '@/components/Marathon/Nav';
 import { AnnouncementList } from '@/components/Marathon/Announcement';
 import { logEvent } from '@/utils/analytics';
 
-/** @todo Replace this with actual data */
-import { announcementItems } from '@/fixtures/marathon/announcements';
+import { announcementItems } from '@/features/marathon';
 
 const Announcements = () => {
   return (
